Object Class for DateItemJPS objects
This object class that does the work of determining the number of inmates that fit our current age range, and or gender / race subset of data selected for a single day in the JailCensusDisplayJPS file. The object class performs this by iterating over the "person" array that is stored at the current index of the "days" array.